INDEV: Separation of duties
Working on getting different concepts seperated out into different module files. Current focus is on database operations since I want that to be mostly streamlined where possible. For now table creation is working just fine but parsing the data from the CSV into the correct forms for sqlite to use automagically is being a bit of a pain. I'll tinker more soon on this.